Rough beginning of decoder.rs

+++ src/instruction.rs
@@ -21,7 +21,7 @@ pub struct Instruction {
     pub command: Op,
     pub segment: Segment,
     pub params: ParameterPair,
-    pub byte_length: u8
+    pub byte_length: u16
 }
 
 impl fmt::Display for Instruction {
